+++ scripts/logtalk_user_setup.js
@@ -3,6 +3,9 @@
 //   This file is part of Logtalk <http://logtalk.org/>  
 //   Copyright (c) 1998-2013 Paulo Moura <pmoura@logtalk.org>
 // 
+//   Logtalk user folder setup script
+//   Last updated on June 3, 2013
+//
 //   This program is free software: you can redistribute it and/or modify
 //   it under the terms of the GNU General Public License as published by
 //   the Free Software Foundation, either version 3 of the License, or
@@ -100,6 +103,13 @@ if (FSObject.FolderExists(logtalk_user)) {
 	if (FSObject.FileExists(logtalk_user + "-backup-" + date + "\\settings.lgt")) {
 		FSObject.CopyFile(logtalk_user + "-backup-" + date + "\\settings.lgt", logtalk_user + "\\settings.lgt");
 		WScript.Echo("Copied your old \"settings.lgt\" file to the new \"\%LOGTALKUSER\%\" directory.");
+		WScript.Echo("The file \"settings-sample.lgt\" file contains a pristine copy of the ");
+		WScript.Echo("\"settings-sample.lgt\" file distributed with the currently installed Logtalk");
+		WScript.Echo("version. Review this file for possible settings files update information.");
+	}
+	if (FSObject.FileExists(logtalk_user + "-backup-" + date + "\\settings.logtalk")) {
+		FSObject.CopyFile(logtalk_user + "-backup-" + date + "\\settings.logtalk", logtalk_user + "\\settings.logtalk");
+		WScript.Echo("Copied your old \"settings.logtalk\" file to the new \"\%LOGTALKUSER\%\" directory.");
 		WScript.Echo("The file \"settings-sample.lgt\" file contains a pristine copy of the ");
 		WScript.Echo("\"settings-sample.lgt\" file distributed with the currently installed Logtalk");
 		WScript.Echo("version. Review this file for possible settings files update information.");